test: move LocalOnly resolution test into finder_test.go
